Updated language configuration
Changed `language` field of `ConfigManager` to
accept a sequence of languages to load to be able to
switch between different languages in the application.

Added field `default_language`, which holds the default
language to use when the application first launches.

Changed `Language` enum to follow ISO 639 two letter
abbreviation standard.