Added state columns to the order_status table
Added columns to determine what state an order status belongs to.
This will make it easier to find all orders that are e.g. "completed",
when there are multiple statuses of that state. By default there are
two statuses for a completed order: "Completed by Technician" and "Completed",
where the latter is assigned when post-processing of the order is completed.

The four possible states are : `is_todo`, `is_in_progress`, `is_on_hold` and
`is_completed`, which are reflected by the boolean columns with the same names.