Added helper methods to BaseDataFrameModel
Added methods:

`get_index` : Get the index of a row from a value of a column.

`get_index_by_row_nr` : Get the index of a row from its row number in the DataFrame.

`get_column` : Get the values of a column in the DataFrame.

Updated `models.core.BaseDataFrameModel` to support defining an int or
string based index column, which reflects the primary key of the database
table being modeled. This gives better type checking capabilities for the
new methods. Added unit tests.