Added permission handling to module app.auth
The IntEnum `Permission` that defines the available permissions, which
are then mapped to the minimum role that has this permission.

Added `Permission.ORDERS_EDIT`, which is mapped to the `UserRole`
meaning that any user with a role of `UserRole` or higher is allowed to
edit orders. The mapping is done through the dictionary `permission_mapping`.

The function `has_permission` is used to check if the current user
has a certain permission.